LayoutTests/picture/expectedfails/picture-img-aria-button-alt-tag.html

<html>
<head>
<script src="../resources/js-test-pre.js"></script>
</head>
<script>
    if (window.testRunner)
        testRunner.dumpAsText();
</script>
<body id="body">

    <picture id='image' src="../resources/cake.png" role="button" alt="alternate" title="title"></picture>

    <p id="description"></p>
    <div id="console"></div>

    <script>
        description("This test makes sure that a picture acting as an aria button still returns its alt tag as the description");

        if (window.accessibilityController) {
            var result = document.getElementById("result");

            var body = document.getElementById("body").focus();
            var imgUIElement = accessibilityController.focusedElement.childAtIndex(0).childAtIndex(0);
            shouldBe("imgUIElement.description", "'AXDescription: alternate'");
        }
    </script>

<script src="../resources/js-test-post.js"></script>
</body>
</html>
